+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/vps/VpsTableColumn.java
@@ -9,6 +9,8 @@ import javafx.geometry.Pos;
 import javafx.scene.control.Label;
 import javafx.scene.control.Tooltip;
 import javafx.scene.layout.HBox;
+import javafx.scene.paint.Paint;
+
 import org.kordamp.ikonli.javafx.FontIcon;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,7 +20,7 @@ import java.util.List;
 public class VpsTableColumn extends HBox {
   private final static Logger LOG = LoggerFactory.getLogger(VpsTableColumn.class);
 
-  public VpsTableColumn(@Nullable String vpsTableId, @Nullable String vpsTableVersionId, @Nullable VPSChanges updates, UISettings uiSettings) {
+  public VpsTableColumn(@Nullable String vpsTableId, @Nullable String vpsTableVersionId, boolean disabled, @Nullable VPSChanges updates, UISettings uiSettings) {
     super(3);
     try {
       int iconSize = 14;
@@ -33,38 +35,38 @@ public class VpsTableColumn extends HBox {
       if (vpsTable != null) {
         vpsTableVersion = vpsTable.getTableVersionById(vpsTableVersionId);
 
-        FontIcon checkboxIcon = WidgetFactory.createCheckboxIcon();
+        FontIcon checkboxIcon = WidgetFactory.createCheckboxIcon(disabled ? WidgetFactory.DISABLED_COLOR : null);
         checkboxIcon.setIconSize(iconSize);
         label.setGraphic(checkboxIcon);
         label.setTooltip(new Tooltip("VPS Table:\n" + vpsTable.getDisplayName()));
       }
       else {
         label.setText(" - ");
-        label.setStyle("-fx-text-fill: #FFFFFF;");
+        label.setStyle(disabled ? WidgetFactory.DISABLED_TEXT_STYLE : WidgetFactory.DEFAULT_TEXT_STYLE);
         label.setTooltip(new Tooltip("No VPS table mapped."));
       }
       this.getChildren().add(label);
 
       label = new Label(" / ");
-      label.setStyle("-fx-text-fill: #FFFFFF;");
+      label.setStyle(disabled ? WidgetFactory.DISABLED_TEXT_STYLE : WidgetFactory.DEFAULT_TEXT_STYLE);
       this.getChildren().add(label);
 
       label = new Label();
       if (vpsTableVersion != null) {
-        FontIcon checkboxIcon = WidgetFactory.createCheckboxIcon();
+        FontIcon checkboxIcon = WidgetFactory.createCheckboxIcon(disabled ? WidgetFactory.DISABLED_COLOR : null);
         checkboxIcon.setIconSize(iconSize);
         label.setGraphic(checkboxIcon);
         label.setTooltip(new Tooltip("VPS Table Version:\n" + vpsTableVersion.toString()));
       }
       else {
         label.setText(" - ");
-        label.setStyle("-fx-text-fill: #FFFFFF;");
+        label.setStyle(disabled ? WidgetFactory.DISABLED_TEXT_STYLE : WidgetFactory.DEFAULT_TEXT_STYLE);
         label.setTooltip(new Tooltip("No VPS table version mapped."));
       }
       this.getChildren().add(label);
 
       label = new Label(" / ");
-      label.setStyle("-fx-text-fill: #FFFFFF;");
+      label.setStyle(disabled ? WidgetFactory.DISABLED_TEXT_STYLE : WidgetFactory.DEFAULT_TEXT_STYLE);
       this.getChildren().add(label);
 
       label = new Label();
@@ -84,6 +86,9 @@ public class VpsTableColumn extends HBox {
 
         if (changeCounter > 0) {
           FontIcon updateIcon = WidgetFactory.createUpdateIcon();
+          if (disabled) {
+            updateIcon.setIconColor(Paint.valueOf(WidgetFactory.DISABLED_COLOR));;
+          }
           label.setGraphic(updateIcon);
 
           String tooltip = "The table or its assets have received updates:\n\n" + builder + "\n\nYou can reset this indicator with the reset action from the context menu.";
@@ -98,7 +103,7 @@ public class VpsTableColumn extends HBox {
       if (changeCounter == 0) {
         label.setText(" - ");
         label.setTooltip(new Tooltip("No updates available."));
-        label.setStyle("-fx-text-fill: #FFFFFF;");
+        label.setStyle(disabled ? WidgetFactory.DISABLED_TEXT_STYLE : WidgetFactory.DEFAULT_TEXT_STYLE);
       }
 
       this.getChildren().add(label);